fix: Home button wrongly detected as Menu button in TSPS (tg5050)

+++ workspace/all/common/api.c
@@ -2613,7 +2613,7 @@ FALLBACK_IMPLEMENTATION void PLAT_pollInput(void) {
 		int pressed = 0; // 0=up,1=down
 		int id = -1;
 		if (event.type == SDL_KEYDOWN || event.type == SDL_KEYUP) {
-			uint8_t code = event.key.keysym.scancode;
+			int code = event.key.keysym.scancode;
 			pressed = event.type == SDL_KEYDOWN;
 			// LOG_info("key event: %i (%i)\n", code,pressed);
 			if (code == CODE_UP) {
@@ -2906,7 +2906,7 @@ FALLBACK_IMPLEMENTATION int PLAT_shouldWake(void) {
 	SDL_Event event;
 	while (SDL_PollEvent(&event)) {
 		if (event.type == SDL_KEYUP) {
-			uint8_t code = event.key.keysym.scancode;
+			int code = event.key.keysym.scancode;
 			if ((BTN_WAKE == BTN_POWER && code == CODE_POWER) || (BTN_WAKE == BTN_MENU && (code == CODE_MENU || code == CODE_MENU_ALT))) {
 				// ignore input while lid is closed
 				if (lid.has_lid && !lid.is_open)
